Clean up Hermod
